new basic assessment and other changes
