+++ action-server/covidflow/actions/action_visit_package.py
@@ -3,9 +3,14 @@ from typing import Any, Dict, List, Text
 from rasa_sdk import Action, Tracker
 from rasa_sdk.executor import CollectingDispatcher
 
-from covidflow.constants import AGE_OVER_65_SLOT, PROVINCE_SLOT, SYMPTOMS_SLOT, Symptoms
+from covidflow.constants import (
+    AGE_OVER_65_SLOT,
+    HAS_CONTACT_RISK_SLOT,
+    PROVINCE_SLOT,
+    SYMPTOMS_SLOT,
+    Symptoms,
+)
 
-from .assessment_form import CONTACT_SLOT, TRAVEL_SLOT
 from .lib.log_util import bind_logger
 
 RISK_LEVEL_MEDICAL = "elevated-medical-risk"
@@ -41,7 +46,7 @@ def _get_risk_level_value(tracker: Tracker) -> str:
         or tracker.get_slot(SYMPTOMS_SLOT) == Symptoms.MODERATE
     ):
         risk_level_value.append(RISK_LEVEL_MEDICAL)
-    if tracker.get_slot(CONTACT_SLOT) is True or tracker.get_slot(TRAVEL_SLOT) is True:
+    if tracker.get_slot(HAS_CONTACT_RISK_SLOT) is True:
         risk_level_value.append(RISK_LEVEL_COVID)
 
     return RISK_LEVEL_DEFAULT if risk_level_value == [] else ",".join(risk_level_value)
